Added verification if result of 'find' is null or not

+++ app/Http/Controllers/ItemController.php
@@ -39,6 +39,10 @@ class ItemController extends Controller
     public function update(Request $request, $id)
     {
         $item = Item::find($id);
+        if(is_null($item))
+        {
+            abort(404);
+        }
         $this->authorize('update', $item);
         $item->done = $request->input('done');
         $item->save();
@@ -54,6 +58,10 @@ class ItemController extends Controller
     public function delete(Request $request, $id)
     {
         $item = Item::find($id);
+        if(is_null($item))
+        {
+            abort(404);
+        }
         $this->authorize('delete', $item);
         $item->delete();
         return $item;
